+++ example/index.ts
@@ -1,21 +1,37 @@
 import { Elysia, t } from 'elysia'
-import z from 'zod'
-import { JSONSchema, Schema } from 'effect'
+import { fromTypes, openapi } from '../src'
+import * as z from 'zod'
 
-import { openapi, withHeaders } from '../src/index'
-
-const app = new Elysia()
+new Elysia()
 	.use(
 		openapi({
-			embedSchema: true,
+			references: fromTypes(),
 			mapJsonSchema: {
 				zod: z.toJSONSchema
 			}
 		})
 	)
-	.get('/test', ({ status }) => status(204, undefined), {
-		response: {
-			204: z.void()
+	.macro('fooBar', {
+		query: z.object({
+			foo: z.optional(z.string())
+		}),
+		resolve({ query }) {
+			return { test: query.foo ? 'foo' : 'bar' }
 		}
 	})
-	.listen(3000)
+	.get(
+		'/',
+		({ test, query }) => {
+			const { foo, bar } = query
+			return { ok: true, test, foo, bar }
+		},
+		{
+			query: z.object({
+				bar: z.optional(z.string())
+			}),
+			fooBar: true
+		}
+	)
+	.listen(3000, () => {
+		console.log('server started')
+	})